overlay: extract live dps calculation to service
