+++ src/main/java/com/duckblade/osrs/dpscalc/plugin/live/TargetedDps.java
@@ -1,0 +1,12 @@
+package com.duckblade.osrs.dpscalc.plugin.live;
+
+import lombok.Value;
+
+@Value
+public class TargetedDps
+{
+
+	private final int npcTarget;
+	private final double dps;
+
+}
