+++ src/main/java/com/duckblade/osrs/dpscalc/plugin/module/DpsPluginModule.java
@@ -1,6 +1,9 @@
 package com.duckblade.osrs.dpscalc.plugin.module;
 
 import com.duckblade.osrs.dpscalc.plugin.DpsMenuActionListener;
+import com.duckblade.osrs.dpscalc.plugin.live.LiveDpsService;
+import com.duckblade.osrs.dpscalc.plugin.live.overlay.LiveDpsOverlay;
+import com.duckblade.osrs.dpscalc.plugin.live.overlay.OverlayMinimizerService;
 import com.duckblade.osrs.dpscalc.plugin.osdata.clientdata.ClientDataProvider;
 import com.duckblade.osrs.dpscalc.plugin.osdata.clientdata.InteractingNpcTracker;
 import com.duckblade.osrs.dpscalc.plugin.osdata.clientdata.RuneLiteClientDataProvider;
@@ -8,8 +11,6 @@ import com.duckblade.osrs.dpscalc.plugin.osdata.wiki.ItemStatsProvider;
 import com.duckblade.osrs.dpscalc.plugin.osdata.wiki.NpcDataProvider;
 import com.duckblade.osrs.dpscalc.plugin.osdata.wiki.WikiItemStatsProvider;
 import com.duckblade.osrs.dpscalc.plugin.osdata.wiki.WikiNpcDataProvider;
-import com.duckblade.osrs.dpscalc.plugin.overlay.LiveDpsOverlay;
-import com.duckblade.osrs.dpscalc.plugin.overlay.OverlayMinimizerService;
 import com.duckblade.osrs.dpscalc.plugin.ui.NavButtonManager;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
@@ -25,6 +26,7 @@ public class DpsPluginModule extends AbstractModule
 		Multibinder<PluginLifecycleComponent> lifecycleComponents = Multibinder.newSetBinder(binder(), PluginLifecycleComponent.class);
 		lifecycleComponents.addBinding().to(DpsMenuActionListener.class);
 		lifecycleComponents.addBinding().to(InteractingNpcTracker.class);
+		lifecycleComponents.addBinding().to(LiveDpsService.class);
 		lifecycleComponents.addBinding().to(LiveDpsOverlay.class);
 		lifecycleComponents.addBinding().to(NavButtonManager.class);
 		lifecycleComponents.addBinding().to(OverlayMinimizerService.class);
