+++ src/main/java/com/duckblade/osrs/dpscalc/plugin/ui/util/ComputeUtil.java
@@ -2,6 +2,7 @@ package com.duckblade.osrs.dpscalc.plugin.ui.util;
 
 import com.duckblade.osrs.dpscalc.calc.exceptions.DpsComputeException;
 import com.duckblade.osrs.dpscalc.calc.exceptions.MissingInputException;
+import java.util.function.Supplier;
 import lombok.experimental.UtilityClass;
 import lombok.extern.slf4j.Slf4j;
 
@@ -22,6 +23,22 @@ public class ComputeUtil
 			{
 				throw e;
 			}
+		}
+	}
+
+	public <T> T tryCompute(Supplier<T> compute)
+	{
+		try
+		{
+			return compute.get();
+		}
+		catch (DpsComputeException e)
+		{
+			if (!(e.getCause() instanceof MissingInputException))
+			{
+				throw e;
+			}
+			return null;
 		}
 	}
